6.0: Move `WebauthnCore::origins_match` to `webauthn-rs-proto`, don't build `tide` docs in CI (#554 cherry-pick) (#559)
* 6.0 cherry-pick: move `WebauthnCore::origins_match` to `webauthn-rs-proto`, don't build `tide` docs in CI (#554)
* Move `origins_match()` to `webauthn-rs-proto`.
`webauthn-rs-core` depends on cryptography libraries, which aren't required
for some configurations of `webauthn-authenticator-rs` (eg: on Windows).
* fix formatting, add proper docs
* Fix clippy lints
* Disable building docs for webauthn-rs-demo in CI
* Fix syntax issue and remove webauthn-rs-demo from manifests
* exclude `tide_tutorial` from docs
* Add dependency on webauthn-rs-core when `ctap2` feature is enabled, fix stubs and missing `use`
