+++ .github/doc_manifest
@@ -17,5 +17,4 @@ webauthn_rs/index.html
 webauthn_rs/interface/index.html
 
 webauthn_rs_core/index.html
-webauthn_rs_demo/index.html
 webauthn_rs_proto/index.html
