+++ .github/workflows/ci.yml
@@ -240,10 +240,17 @@ jobs:
       # This tests that all the stubs work properly for optional dependencies,
       # but doesn't work for fido-key-manager which includes NFC and USB support
       # by default.
+      #
+      # tide_tutorial and webauthn-rs-demo are excluded from doc builds because
+      # of https://github.com/bytecodealliance/rustix/issues/1620, which is
+      # broken on Rust nightly.
       - name: generate docs
         run: |
-          cargo ${{ matrix.rust_version == 'nightly' && '+nightly' || '' }} doc --all \
+          cargo ${{ matrix.rust_version == 'nightly' && '+nightly' || '' }} doc \
+            --all \
             --exclude fido-key-manager \
+            --exclude tide_tutorial \
+            --exclude webauthn-rs-demo \
             --no-deps \
             --document-private-items
         env:
@@ -263,10 +270,12 @@ jobs:
           sudo apt-get -y install libpcsclite-dev libusb-1.0-0-dev libdbus-1-dev
       - run: |
           cargo ${{ matrix.rust_version == 'nightly' && '+nightly' || '' }} doc \
-          --all \
-          --no-deps \
-          --document-private-items \
-          --all-features
+            --all \
+            --exclude tide_tutorial \
+            --exclude webauthn-rs-demo \
+            --no-deps \
+            --document-private-items \
+            --all-features
         env:
           RUSTDOCFLAGS: ${{ matrix.rust_version == 'nightly' && '--cfg docsrs' || '' }}
       - uses: actions/upload-artifact@v4
