+++ webauthn-authenticator-rs/Cargo.toml
@@ -50,6 +50,7 @@ ctap2 = [
     "dep:serde_bytes",
     "dep:tokio",
     "dep:tokio-stream",
+    "dep:webauthn-rs-core",
 ]
 ctap2-management = ["ctap2"]
 # Support for SoloKey's vendor commands
@@ -71,8 +72,8 @@ rustdoc-args = ["--cfg", "docsrs"]
 
 [dependencies]
 base64.workspace = true
-webauthn-rs-proto.workspace = true
-webauthn-rs-core = { workspace = true }
+webauthn-rs-proto = { workspace = true, features = ["origin"] }
+webauthn-rs-core = { workspace = true, optional = true }
 crypto-glue.workspace = true
 
 tracing.workspace = true
